requirements.txt
asgiref==3.8.1
bleach==6.2.0
coverage==7.8.0
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
redis==6.0.0
sqlparse==0.5.3
typing_extensions==4.13.2
webencodings==0.5.1
whitenoise==6.9.0
